fix typo

+++ orchagent/fabricportsorch.cpp
@@ -1098,7 +1098,7 @@ void FabricPortsOrch::updateFabricRate()
         {
             const auto field = fvField(fv);
             const auto value = fvValue(fv);
-            for (int cnt = 0; cnt != cntNames.size(); cnt++)
+            for (size_t cnt = 0; cnt != cntNames.size(); cnt++)
             {
                 if (field == "SAI_PORT_STAT_IF_OUT_OCTETS")
                 {
